extract network communication into server_packets.py
