+++ src/coop/_coopwidget.py
@@ -67,8 +67,8 @@ class CoopWidget(FormClass, BaseClass):
         self.options = []
 
         self.client.showCoop.connect(self.coopChanged)
-        self.client.coopInfo.connect(self.processCoopInfo)
-        self.client.gameInfo.connect(self.processGameInfo)
+        self.client.net.coopInfo.connect(self.processCoopInfo)
+        self.client.net.gameInfo.connect(self.processGameInfo)
         self.coopList.header().setResizeMode(0, QtGui.QHeaderView.ResizeToContents)
         self.coopList.setItemDelegate(CoopMapItemDelegate(self))
 
@@ -79,7 +79,7 @@ class CoopWidget(FormClass, BaseClass):
         self.coopList.itemDoubleClicked.connect(self.coopListDoubleClicked)
         self.coopList.itemClicked.connect(self.coopListClicked)
 
-        self.client.coopLeaderBoard.connect(self.processLeaderBoardInfos)
+        self.client.net.coopLeaderBoard.connect(self.processLeaderBoardInfos)
         self.tabLeaderWidget.currentChanged.connect(self.askLeaderBoard)
 
         self.linkButton.clicked.connect(self.linkVanilla)
